create ObjectMapperConfig, delete JsonFormat
