+++ src/main/java/faang/school/notificationservice/dto/comment/CommentEvent.java
@@ -1,6 +1,5 @@
 package faang.school.notificationservice.dto.comment;
 
-import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
@@ -12,6 +11,5 @@ public class CommentEvent {
     private long postId;
     private long authorId;
     private long commentId;
-    @JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss.SSSSSS")
     private LocalDateTime timestamp;
 }
