+++ src/main/resources/application.yaml
@@ -41,6 +41,7 @@ logging:
   level:
     root: info
 
+#Можно зарегистрировать аккаунт без привязки карты. Будет 2 евро на отправку тестовых SMS
 vonage: # https://dashboard.nexmo.com/getting-started/sms
   api:
     key: whatever
